reformat text modules
