+++ pvops/tests/test_text.py
@@ -11,15 +11,20 @@ import sys
 text_directory = os.path.join("..", "text")
 sys.path.append(text_directory)
 
-from text_remove_nondate_nums import text_remove_nondate_nums
-from text_remove_numbers_stopwords import text_remove_numbers_stopwords
-from create_stopwords import create_stopwords
-from get_dates import get_dates
-from visualize_attribute_timeseries import visualize_attribute_timeseries
-from visualize_word_frequency_plot import visualize_word_frequency_plot
-from visualize_attribute_connectivity import visualize_attribute_connectivity
-from summarize_text_data import summarize_text_data
+# from text_remove_nondate_nums import text_remove_nondate_nums
+# from text_remove_numbers_stopwords import text_remove_numbers_stopwords
+# from create_stopwords import create_stopwords
+# from get_dates import get_dates
+# from visualize_attribute_timeseries import visualize_attribute_timeseries
+# from visualize_word_frequency_plot import visualize_word_frequency_plot
+# from visualize_attribute_connectivity import visualize_attribute_connectivity
+# from summarize_text_data import summarize_text_data
 
+from classify import *
+from nlp_utils import *
+from preprocess import *
+from utils import *
+from visualize import *
 
 def test_text_remove_nondate_nums():
     example = r"This is a test example https://www.google.com 10% #10 101 1-1-1 a-e4 13-1010 10.1 123456789 123/12 executed on 2/4/2020"
